[FIX] fix multi-language for website.seo fields and some refactoring for ir_translation and res_lang (#12)
