+++ website_seo/models/__init__.py
@@ -21,4 +21,5 @@
 from . import ir_http
 from . import ir_ui_view
 from . import ir_translation
+from . import res_lang
 from . import website
